requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
numpy==2.4.3
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.1
